# docs/swagger.yaml
openapi: 3.0.3
info:
  title: MindFlow API
  version: 1.0.0
  description: Comprehensive API documentation for the Mental Health Trend Analytics Platform.
servers:
  - url: http://localhost:5000
...
paths:
...
  /api/users:
    get:
      tags: [Users]
      summary: Get all users for provider dashboard
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User list returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserListItem'
        '403':
          description: Forbidden for patient role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
  /api/users/create:
    post:
      tags: [Users]
      summary: Create a patient profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Patient created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'
        '400':
          description: Validation or duplicate user failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden for patient role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
  /api/users/{id}/details:
    get:
      tags: [Users]
      summary: Get detailed patient information
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Patient detail returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '404':
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users/profile:
    put:
      tags: [Users]
      summary: Update the logged-in user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users/password:
    put:
      tags: [Users]
      summary: Change the current user password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid current password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users/settings:
    put:
      tags: [Users]
      summary: Update current user settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSettingsRequest'
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: string
    ForbiddenResponse:
      type: object
      properties:
        message:
          type: string
        alert:
          type: string
    Address:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        country:
          type: string
    Settings:
      type: object
      properties:
        notifications:
          type: object
          properties:
            email:
              type: boolean
            push:
              type: boolean
            weeklyReports:
              type: boolean
            patientAlerts:
              type: boolean
            systemUpdates:
              type: boolean
        appearance:
          type: object
          properties:
            theme:
              type: string
            colorScheme:
              type: string
            fontSize:
              type: string
...
    AuthResponse:
      type: object
      properties:
        _id:
          type: string
        username:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        bio:
          type: string
        profileImage:
          type: string
          nullable: true
        settings:
          $ref: '#/components/schemas/Settings'
        role:
          type: string
        token:
          type: string
...
    CreateUserRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        dateOfBirth:
          type: string
          format: date
        sex:
          type: string
        phone:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        bloodType:
          type: string
        allergies:
          type: array
          items:
            type: string
        medicalHistory:
          type: array
          items:
            type: string
        profileImage:
          type: string
        mentalHealth:
          type: object
    CreateUserResponse:
      type: object
      properties:
        _id:
          type: string
        patientId:
          type: string
        username:
          type: string
        email:
          type: string
        age:
          type: integer
          nullable: true
        sex:
          type: string
        profileImage:
          type: string
          nullable: true
        role:
          type: string
    UserListItem:
      allOf:
        - $ref: '#/components/schemas/CreateUserResponse'
        - type: object
          properties:
            lastCheckIn:
              type: string
              format: date-time
              nullable: true
            status:
              type: string
    UserDetail:
      type: object
      properties:
        _id:
          type: string
        patientId:
          type: string
        username:
          type: string
        email:
          type: string
        role:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        age:
          type: integer
        sex:
          type: string
        phone:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        bloodType:
          type: string
        allergies:
          type: array
          items:
            type: string
        medicalHistory:
          type: array
          items:
            type: string
        mentalHealth:
          type: object
        bio:
          type: string
        profileImage:
          type: string
          nullable: true
        settings:
          $ref: '#/components/schemas/Settings'
        createdAt:
          type: string
          format: date-time
    UpdateProfileRequest:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        bio:
          type: string
        profileImage:
          type: string
    UpdateSettingsRequest:
      type: object
      properties:
        notifications:
          type: object
        appearance:
          type: object
...